# api/openapi.yaml
openapi: 3.1.0
info:
  title: Bloom API
  version: 0.1.0
  summary: Self-hosted Jellyfin companion; invites, statistics, and requests.
  description: |
    Bloom's JSON API. It is the only way the embedded SPA (ADR 0002) talks to
    the backend, so this document is the contract the frontend builds against.

    Conventions (from the Go handbook, foundations/serialization.md):

    - Every error response, on every endpoint, is the `ErrorResponse` envelope.
      Unknown routes under `/api/` return it too, never HTML.
    - `code` is a machine-readable string enum the client may branch on. It is
      not the HTTP status. The set only grows; a code is never repurposed.
    - 5xx bodies are opaque: a generic `message` plus `request_id`. The detail
      lives in the server log under that id.
    - Every response carries `X-Request-ID`. A client may send its own (up to
      128 characters) and it is echoed; otherwise one is generated.
    - Field names are `snake_case`. Timestamps are RFC 3339 in UTC.
    - Browser authentication uses a server-side session. The cookie contains
      only an opaque session token and is never an API credential.
    - Permission identifiers are stable API contract values. The catalog may
      grow additively; published identifiers are never renamed or removed.
servers:
  - url: /
    description: Same origin as the SPA.
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Start a local browser session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated. Sets the `bloom_session` cookie.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAccountResponse"
        "401":
          description: Invalid credentials, or a malformed inbound session. Credential failures are indistinguishable. A rejected session expires the cookie with `Set-Cookie` before credential authentication runs.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cross-origin request rejected by CSRF protection.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Missing or unsupported request media type. The `code` is `unsupported_media_type`.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Malformed JSON or invalid fields.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Login rate limit exceeded. `Retry-After` is seconds.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Retry-After: { $ref: "#/components/headers/RetryAfter" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/AuthInternalError"
        "405":
          $ref: "#/components/responses/AuthMethodNotAllowedPost"
        "503":
          description: Password-verification capacity is exhausted. `Retry-After` is seconds.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Retry-After: { $ref: "#/components/headers/RetryAfter" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Destroy the current browser session.
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Session destroyed.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
        "401":
          description: No live session.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Cross-origin request rejected by CSRF protection.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/AuthInternalError"
        "405":
          $ref: "#/components/responses/AuthMethodNotAllowedPost"
  /api/v1/auth/me:
    get:
      tags: [auth]
      operationId: getCurrentAccount
      summary: Return the account for the current browser session.
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Current account.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentAccountResponse"
        "401":
          description: No live session, or the account was deleted or disabled.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/AuthInternalError"
        "405":
          $ref: "#/components/responses/AuthMethodNotAllowedGet"
  /api/v1/auth/permissions:
    get:
      tags: [auth]
      operationId: listPermissions
      summary: Return the public stable permission catalog.
      description: Permission identifiers only grow additively and are never renamed or removed.
      responses:
        "200":
          description: The complete permission catalog in stable contract order.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlPublic" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PermissionCatalogResponse"
        "405":
          $ref: "#/components/responses/AuthMethodNotAllowedGet"
  /api/v1/roles:
    get:
      tags: [roles]
      operationId: listRoles
      summary: List roles and their effective permission sets.
      security:
        - sessionCookie: []
      parameters:
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned as `next_cursor` by the previous page.
          schema:
            type: string
            minLength: 1
            maxLength: 86
        - name: page_size
          in: query
          required: false
          description: Requested role count. Defaults to 50; values above 100 are clamped to 100.
          schema:
            type: integer
            minimum: 1
            default: 50
      responses:
        "200":
          description: At most 100 roles ordered by name, with each permission set sorted.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RolesResponse"
        "401":
          description: No live session. Invalid or revoked session cookies are expired.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The account lacks `admin.roles`. The `code` is `forbidden`.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Invalid cursor or page-size query value.
          headers:
            Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
            Set-Cookie: { $ref: "#/components/headers/SetCookie" }
            Vary: { $ref: "#/components/headers/VaryCookie" }
            X-Request-ID: { $ref: "#/components/headers/XRequestID" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/AuthInternalError"
        "405":
          $ref: "#/components/responses/AuthMethodNotAllowedGet"
  /api/v1/media-servers:
    post:
      tags: [media-servers]
      operationId: createMediaServer
      summary: Probe and register a media server.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaServerRequest"
      responses:
        "201":
          $ref: "#/components/responses/MediaServerCreated"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/MediaServerFailure" }
        "503": { $ref: "#/components/responses/MediaServerBusy" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedPostGet" }
    get:
      tags: [media-servers]
      operationId: listMediaServers
      summary: List registered media servers.
      security:
        - sessionCookie: []
      parameters:
        - name: cursor
          in: query
          schema: { type: string, minLength: 1, maxLength: 400 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, default: 50 }
      responses:
        "200":
          $ref: "#/components/responses/MediaServersListed"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedPostGet" }
  /api/v1/media-servers/{id}:
    get:
      tags: [media-servers]
      operationId: getMediaServer
      summary: Return one registered server and its capabilities.
      parameters:
        - $ref: "#/components/parameters/MediaServerID"
      security:
        - sessionCookie: []
      responses:
        "200":
          $ref: "#/components/responses/MediaServerReturned"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedGetDelete" }
    delete:
      tags: [media-servers]
      operationId: deleteMediaServer
      summary: Delete a registered media server.
      parameters:
        - $ref: "#/components/parameters/MediaServerID"
      security:
        - sessionCookie: []
      responses:
        "204": { $ref: "#/components/responses/MediaServerDeleted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedGetDelete" }
  /api/v1/media-servers/{id}/probe:
    post:
      tags: [media-servers]
      operationId: probeMediaServer
      summary: Re-probe a registered media server.
      parameters:
        - $ref: "#/components/parameters/MediaServerID"
      security:
        - sessionCookie: []
      responses:
        "200":
          $ref: "#/components/responses/MediaServerProbed"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/MediaServerFailure" }
        "503": { $ref: "#/components/responses/MediaServerBusy" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedPost" }
  /api/v1/media-servers/{id}/libraries:
    get:
      tags: [media-servers]
      operationId: listMediaServerLibraries
      summary: List libraries visible to a registered media server credential.
      parameters:
        - $ref: "#/components/parameters/MediaServerID"
      security:
        - sessionCookie: []
      responses:
        "200":
          $ref: "#/components/responses/MediaServerLibrariesListed"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
        "502": { $ref: "#/components/responses/MediaServerFailure" }
        "503": { $ref: "#/components/responses/MediaServerBusy" }
        "405": { $ref: "#/components/responses/AuthMethodNotAllowedGet" }
components:
...
  headers:
    CacheControlNoStore:
      description: Prevents storage of a response that touched browser-session state.
      schema:
        type: string
        enum: [no-store]
    CacheControlPublic:
      description: Allows shared caching of the public additive catalog for one hour.
      schema:
        type: string
        const: public, max-age=3600
    VaryCookie:
      description: Marks the response as varying on the inbound Cookie header.
      schema:
        type: string
        enum: [Cookie]
    XRequestID:
      description: Correlation identifier generated or echoed by Bloom.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    SetCookie:
      description: Creates, refreshes, or expires the `bloom_session` cookie when applicable.
      schema:
        type: string
    RetryAfter:
      description: Whole seconds until the next attempt may be made. Media-server values are capped at 30 seconds.
      schema:
        type: integer
        minimum: 1
  responses:
    MediaServerCreated:
      description: The connection was proved and stored with its credential encrypted.
      headers:
        Cache-Control: { $ref: "#/components/headers/CacheControlNoStore" }
        Set-Cookie: { $ref: "#/components/headers/SetCookie" }
        Vary: { $ref: "#/components/headers/VaryCookie" }
        X-Request-ID: { $ref: "#/components/headers/XRequestID" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/CreateMediaServerResponse" }
...
    AuthMethodNotAllowedPost:
      description: The route accepts only `POST`.
      headers:
        Allow:
          description: Allowed method for this route.
          schema:
            type: string
            const: POST
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthMethodNotAllowedGet:
      description: The route accepts only `GET`.
      headers:
        Allow:
          description: Allowed method for this route.
          schema:
            type: string
            const: GET
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthInternalError:
      description: Opaque authentication or session failure. `Set-Cookie` may refresh or expire the session.
      headers:
        Cache-Control:
          $ref: "#/components/headers/CacheControlNoStore"
        Set-Cookie:
          $ref: "#/components/headers/SetCookie"
        Vary:
          $ref: "#/components/headers/VaryCookie"
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: bloom_session
  schemas:
...
    CreateMediaServerRequest:
      type: object
      required: [kind, name, base_url, api_key]
      additionalProperties: false
      properties:
        kind: { $ref: "#/components/schemas/MediaServerKind" }
        name:
          type: string
          minLength: 1
          x-max-bytes: 100
          description: Length is measured in UTF-8 bytes and must not exceed 100.
        base_url:
          type: string
          format: uri
          pattern: "^https?://"
          x-max-bytes: 2048
          description: Length is measured in UTF-8 bytes and must not exceed 2048. HTTPS is required unless allow_insecure is true.
        api_key:
          type: string
          format: password
          minLength: 1
          x-max-bytes: 4096
          description: Length is measured in UTF-8 bytes and must not exceed 4096. Quotes and backslashes are accepted; control characters are not.
          writeOnly: true
        allow_insecure:
          type: boolean
          default: false
          description: Explicitly permits plaintext HTTP and its credential-exposure risk. It must remain false for HTTPS.
...
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 256
          description: Raw submitted username, bounded before canonicalization. The server canonicalizes it with the PRECIS UsernameCaseMapped profile, then validates the canonical comparison key's length and syntax. Equivalent case, width, and normalized forms are accepted.
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 1024
          description: Length is measured in Unicode code points.
...
    CurrentAccountResponse:
      type: object
      required: [account, roles, permissions]
      additionalProperties: false
      properties:
        account:
          $ref: "#/components/schemas/AccountSummary"
        roles:
          type: array
          items:
            type: string
          description: Assigned role names in sorted order.
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
          description: Effective permissions in sorted order.
...
    PermissionCatalogResponse:
      type: object
      required: [permissions]
      additionalProperties: false
      properties:
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionCatalogEntry"
...
    RolesResponse:
      type: object
      required: [items, next_cursor]
      additionalProperties: false
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/Role"
        next_cursor:
          type: string
          maxLength: 86
          description: Opaque cursor for the next page, or an empty string on the final page.
...
    ErrorResponse:
      type: object
      description: The single structured error envelope returned by every endpoint.
      required: [code, message, request_id]
      additionalProperties: false
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable and safe to surface. Generic for 5xx.
        fields:
          type: array
          description: Per-field validation failures. Present only when `code` is `validation_failed`.
          items:
            $ref: "#/components/schemas/FieldError"
        request_id:
          type: string
          description: Correlation id, also sent as the `X-Request-ID` header.
...